requirements.txt
aniso8601==9.0.1
async-timeout==4.0.2
click==8.1.3
colorama==0.4.5
confluent-kafka==1.9.2
Deprecated==1.2.13
Flask==2.2.2
Flask-Login==0.6.2
Flask-RESTful==0.3.8
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
m-caching==0.1.8
m-kafka-sdk-v2==0.1.4.2
m-schedule==0.6.7
m-singleton==0.3
m-threadpool==0.2
m-validator==0.1
MarkupSafe==2.1.1
packaging==21.3
PyJWT==1.7.1
pymongo==4.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
redis==4.3.4
schedule==1.1.0
six==1.16.0
Unidecode==1.3.6
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.8.1
